% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCTable.R
\name{aucfromdor}
\alias{aucfromdor}
\title{Estimate AUC from a Diagnostic Odds Ratio}
\usage{
aucfromdor(dor)
}
\arguments{
\item{dor}{A Diagnostic Odds Ratio}
}
\value{
auc, the area under the theoretical proper ROC curve
}
\description{
Estimate the Area Under a ROC curve (AUC) under the
assumption of a proper model using the corresponding Diagnostic
Odds Ratio (the parameter of the theoretical ROC curve with the
same AUC). In a proper ROC curve DOR must be >= 1.0.
}
\examples{
aucfromdor(1.0)
# AUC = 0.5 corresponds to a non informative ROC curve

aucfromdor(8.0)
# A Diagnostic Odds Ratio >= 0.8 is needed to obtain an AUC of at least 0.8

aucfromdor(0.9)
# A Diagnostic Odds Ratio > 1.0 corresponds to a non-proper ROC curve

}
