% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCProper.R
\name{biasproproc}
\alias{biasproproc}
\title{Bias in Using AUC of a Proper ROC Curve for Diagnostic Accuracy}
\usage{
biasproproc(dor)
}
\arguments{
\item{dor}{Diagnostic Odds Ratio (the parameter of the expected proper curve)}
}
\value{
the bias, the difference between AUC and the "best" expected accuracy
}
\description{
Estimate the difference between AUC and the global
diagnostic accuracy, i.e., the proportion of corrected classifications
expected at the best cut-off of a proper ROC curve in a Proper ROC Model
}
\examples{
ROCCa125 <- roctable(PancreaticData$Ca125, PancreaticData$Status)
BiasCa125 <- biasproproc(rocdor(rocauc(ROCCa125)))
print(paste("Bias for the proper ROC curve for Ca125 tumor marker: ",
      round(BiasCa125,4)))


}
