% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OvarianData.R
\docType{data}
\name{OvarianData}
\alias{OvarianData}
\title{Ovarian Cancer Data Set (First Two Genes)}
\format{A data matrix}
\usage{
OvarianData
}
\description{
Expression of two genes out of more than 1500 genes
in normal (n = 23) and malignant (n = 30) ovarian tissue.
Data were extracted from the publicly available data set of
the Diagnostic and Biomarkers Statistical (DABS) Center at the
Fred Hutchinson Cancer Research Center:
http://research.fhcrc.org/diagnostic-biomarkers-center/en/datasets.html
}
\keyword{datasets}
