% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PancreaticData.R
\docType{data}
\name{PancreaticData}
\alias{PancreaticData}
\title{Pancreatic Cancer Data Set}
\format{A data matrix}
\usage{
PancreaticData
}
\description{
The data set includes 141 subjects (51 controls and 90 cases)
and two tumour markers (namely: Ca19.9 and Ca125) in the peripheral blood.
Data were extracted from the publicly available data set of the
Diagnostic and Biomarkers Statistical (DABS) Center at the
Fred Hutchinson Cancer Research Center:
http://research.fhcrc.org/diagnostic-biomarkers-center/en/datasets.html
}
\keyword{datasets}
