% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCProper.R
\name{parpropauc}
\alias{parpropauc}
\title{Partial Proper AUC}
\usage{
parpropauc(dor, x1, x2)
}
\arguments{
\item{dor}{Diagnostic Odds Ratio (the parameter of a proper ROC curve)}

\item{x1}{First value of 1 - specificity}

\item{x2}{second First value of 1 - specificity}
}
\value{
pauc, the partial area under the theoretical proper ROC curve
}
\description{
Calculate the partial AUC between two values of 1-specificity
in a theoretical proper ROC curve
(for internal use only: called by tprc()).
}
