% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCPlot.R
\name{plotproper}
\alias{plotproper}
\title{Add a Theoretical Proper ROC Curve to a ROC Plot}
\usage{
plotproper(dor, color = "black")
}
\arguments{
\item{dor}{The Diagnostic Odds Ratio of the proper curve}

\item{color}{The color of the theoretical proper ROC curve}
}
\description{
Add a theoretical proper ROC curve to a pre-existing ROC plot
generated by plotroc()
}
\examples{
ROCGene2 <- roctable(OvarianData$Gene2, OvarianData$Status)
plotroc(ROCGene2)
plotproper(rocdor(rocauc(ROCGene2)))

}
