% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCPlot.R
\name{plotroc}
\alias{plotroc}
\title{Plot One or More ROC Curves from the Corresponding ROC Tables (S3 Objects)}
\usage{
plotroc(roctable1, roctable2 = NULL, roctable3 = NULL)
}
\arguments{
\item{roctable1}{A ROC Table (S3 object)}

\item{roctable2}{A ROC Table (S3 object)}

\item{roctable3}{A ROC Table (S3 object)}
}
\description{
Plot up to three empirical ROC curves from the corresponding
ROC tables. A ROC table is an S3 object derived from an R matrix, generated
by the roctable() function
}
\examples{
ROCGene1 <- roctable(OvarianData$Gene1, OvarianData$Status)
ROCGene2 <- roctable(OvarianData$Gene2, OvarianData$Status)
plotroc(ROCGene1, ROCGene2)

}
