% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCPlot.R
\name{proproc}
\alias{proproc}
\title{Calculate the Coordinates of a Proper ROC Curve (internal usage)}
\usage{
proproc(dor, npoints = 2001L)
}
\arguments{
\item{dor}{The Diagnostic Odds Ratio of the proper curve}

\item{npoints}{Number of points to be plotted}
}
\description{
Calculate the coordinates of a proper ROC curve
from the corresponding Diagnostic Odds Ratio (DOR). Called by
plotproper(), internal usage only
}
