% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCProper.R
\name{pvaltprc}
\alias{pvaltprc}
\title{p-value for the Test for a Proper ROC Curve}
\usage{
pvaltprc(roctab, numiter = 200L)
}
\arguments{
\item{roctab}{A ROC Table (S3 object derived from a S3 matrix)}

\item{numiter}{The number of Monte Carlo simulations}
}
\value{
pval, the p-value for the Test for a Proper ROC Curve (TPRC)
}
\description{
Calculate the p-value for the TPRC test statistic. Test the
departure from a proper ROC model using a Monte Carlo simulation.
The test might take some minutes to be completed
}
\examples{
ROCGene1 <- roctable(OvarianData$Gene1, OvarianData$Status)
pvalROCGene1 <- pvaltprc(ROCGene1)
print(paste("Test for a Proper ROC curve for Gene1: p = ",
      round(pvalROCGene1,4)))

}
