% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCTable.R
\name{rocauc}
\alias{rocauc}
\title{Estimate AUC from a ROC Table}
\usage{
rocauc(roc)
}
\arguments{
\item{roc}{A ROC Table (S3 object)}
}
\value{
auc, the area under the ROC curve whose coordinates are in the
first two columns of the ROC table
}
\description{
Calculate the area under a ROC curve from a ROC Table.
ROC Table is an S3 object derived from an R matrix and generated by
the roctable() function
}
\examples{
ROCCa125 <- roctable(PancreaticData$Ca125, PancreaticData$Status)
AUCCa125 <- rocauc(ROCCa125)
print(paste("AUC for Ca125 tumour marker: ", round(AUCCa125,4)))


}
