% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCTable.R
\name{roccoord}
\alias{roccoord}
\title{Calculate the Coordinates of a ROC curve}
\usage{
roccoord(marker, status)
}
\arguments{
\item{marker}{The disease marker (a continuous variable)}

\item{status}{Patient status: 0 = control,  1 = case}
}
\value{
matrix, a matrix containing values of sens. and 1-spec.
}
\description{
Calculate the coordinates of a ROC curve
(called by: pvaltprc(). For internal use only)
}
