% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCTable.R
\name{rocdor}
\alias{rocdor}
\title{Estimate DOR from an AUC under a Proper Model Assumption}
\usage{
rocdor(auc, eps = 1e-04, MAXITER = 1000L)
}
\arguments{
\item{auc}{The area under a ROC curve}

\item{eps}{The precision required for DOR estimate (default: 0.0001)}

\item{MAXITER}{the maximum number of iterations for the DOR estimate
(default: 1000)}
}
\value{
dor, the Diagnostic Odds Ratio (the parameter defining a proper
ROC curve)
}
\description{
Estimate the Diagnostic Odds Ratio (DOR) corresponding
to the area under a ROC curve (AUC), under the assumption of a proper
model.
}
\examples{
ROCCa125 <- roctable(PancreaticData$Ca125, PancreaticData$Status)
DORCa125 <- rocdor(rocauc(ROCCa125))
print(paste("DOR for the proper ROC curve for Ca125 tumor marker: ",
      round(DORCa125,4)))

}
