% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCTable.R
\name{roctable}
\alias{roctable}
\title{Make a ROC Table}
\usage{
roctable(marker, status)
}
\arguments{
\item{marker}{The disease marker (a continuous variable)}

\item{status}{Patient status: 0 = control,  1 = case}
}
\value{
rocmatr, the ROC Table Matrix (S3 object)
}
\description{
Make a ROC Table (S3 object derived from a R matrix).
The ROC Table includes seven columns, corresponding to the following
variables: sensitivity ("1-Sens"), 1 - specificity ("1-Spec"),
Youden index ("J"), Global Diagnostic Accuracy ("GDA"), Positive
Likelihood Ratio ("LR+"), Negative Likelihood Ratio ("LR-"), and
the corresponding cut-off values ("Cut-off")
}
\examples{
## Not Run:
RocCa125 <- roctable(PancreaticData$Ca125, PancreaticData$Status)
head(RocCa125)

}
