% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCProper.R
\name{tprc}
\alias{tprc}
\title{Test for a Proper ROC Curve}
\usage{
tprc(roctab, dor = -1)
}
\arguments{
\item{roctab}{A ROC Table (S3 object derived from a S3 matrix)}

\item{dor}{The Diagnostic Odds Ratio that defines the proper ROC curve}
}
\value{
tprc, the difference between the empirical and the theoretical AUC
under a proper model assumption
}
\description{
Calculate the test statistic (TPRC) for the departure from a
proper ROC model. TPRC is obtained as the difference between the observed
(empirical) AUC and the corresponding AUC obtained under a proper model
}
